telegram-stock-bot/requirements.txt
python-telegram-bot[job-queue]==21.6
yfinance==0.2.54
pandas==2.2.3
numpy==2.1.1
python-dotenv==1.0.1
tzdata==2025.2
